Add an event listener to the board display
Every div is given an event listener which changes its color while the
mouse cursor is hovering over it and the primary mouse button is held
down. This allows the user to draw on the board.

Also added some auxiliary code to check if the mouse button is currently
being held down.