Add createNewBoard() and clearBoard() functions
Added a clearBoard() function which removes the display board altogether
and replaces it with a new empty <div> with the right id.
createNewBoard() prompts the user for the desired size of a new board
and generates one.

Also added a button which allows the user to call createNewBoard() at
will.